feat: add lenis for smooth scroll experience
