+++ web-next/package.json
@@ -39,6 +39,7 @@
     "embla-carousel-react": "^8.6.0",
     "input-otp": "^1.4.2",
     "katex": "^0.16.22",
+    "lenis": "^1.3.8",
     "lexical": "^0.33.1",
     "lodash-es": "^4.17.21",
     "lucide-react": "^0.525.0",
