--- web-next/src/components/app-provider.tsx
@@ -1,20 +1,0 @@
-"use client";
-
-import { ThemeProvider } from "@/components/theme-provider";
-import { Toaster } from "@/components/ui/sonner";
-
-// https://ui.shadcn.com/docs/dark-mode/next#wrap-your-root-layout
-
-export function AppProvider({ children }: { children: React.ReactNode }) {
-  return (
-    <ThemeProvider
-      attribute="class"
-      defaultTheme="system"
-      enableSystem
-      disableTransitionOnChange
-    >
-      {children}
-      <Toaster />
-    </ThemeProvider>
-  );
-}
