+++ web-next/src/components/providers/theme-provider.tsx
@@ -3,8 +3,10 @@
 import { ThemeProvider as NextThemesProvider } from "next-themes";
 import * as React from "react";
 
-// https://ui.shadcn.com/docs/dark-mode/next#create-a-theme-provider
-
+/**
+ * @description Theme provider from shadcn/ui
+ * @see https://ui.shadcn.com/docs/dark-mode/next#create-a-theme-provider
+ */
 export function ThemeProvider({
   children,
   ...props
